fix and tidy up
