+++ Assets/Scenes/Main.unity
@@ -162,9 +162,7 @@ GameObject:
   - component: {fileID: 512232610}
   - component: {fileID: 512232609}
   - component: {fileID: 512232608}
-  - component: {fileID: 512232611}
   - component: {fileID: 512232612}
-  - component: {fileID: 512232613}
   m_Layer: 0
   m_Name: Main Camera
   m_TagString: MainCamera
@@ -237,26 +235,6 @@ Transform:
   m_Father: {fileID: 1457566343}
   m_RootOrder: 0
   m_LocalEulerAnglesHint: {x: 45, y: 0, z: 0}
---- !u!114 &512232611
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 512232607}
-  m_Enabled: 0
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 0c4fb070738f3544ea560ef08186d72d, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  panLimit1: {x: 0, y: -120}
-  panLimit2: {x: 120, y: 50}
-  minY: 40
-  maxY: 150
-  scrollSpeed: 10
-  panSpeed: 60
-  dragStartPosition: {x: 0, y: 0, z: 0}
-  dragCurrentPosition: {x: 0, y: 0, z: 0}
 --- !u!114 &512232612
 MonoBehaviour:
   m_ObjectHideFlags: 0
@@ -270,36 +248,6 @@ MonoBehaviour:
   m_Name: 
   m_EditorClassIdentifier: 
   targetsTag: Tile
---- !u!114 &512232613
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 512232607}
-  m_Enabled: 0
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 16d7ee547532bfc4cbaf9f7ebeacead1, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  cameraTrasnform: {fileID: 512232610}
-  panLimit1: {x: 0, y: -120}
-  panLimit2: {x: 120, y: 50}
-  minY: 40
-  maxY: 200
-  normalSpeed: 5
-  fastSpeed: 12
-  movementSpeed: 5
-  movementTime: 5
-  rotationAmount: 5
-  zoomAmount: {x: 0, y: 0, z: 0}
-  newPosition: {x: 0, y: 0, z: 0}
-  newRotation: {x: 0, y: 0, z: 0, w: 0}
-  newZoom: {x: 0, y: 0, z: 0}
-  dragStartPosition: {x: 0, y: 0, z: 0}
-  dragCurrentPosition: {x: 0, y: 0, z: 0}
-  rotateStartPosition: {x: 0, y: 0, z: 0}
-  rotateCurrentPosition: {x: 0, y: 0, z: 0}
 --- !u!1 &834136085
 GameObject:
   m_ObjectHideFlags: 0
@@ -324,19 +272,23 @@ MonoBehaviour:
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
   m_GameObject: {fileID: 834136085}
-  m_Enabled: 0
+  m_Enabled: 1
   m_EditorHideFlags: 0
   m_Script: {fileID: 11500000, guid: 0c4fb070738f3544ea560ef08186d72d, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  panLimit1: {x: 0, y: 0}
-  panLimit2: {x: 0, y: 0}
-  minY: 20
-  maxY: 150
-  scrollSpeed: 2
-  panSpeed: 30
-  dragStartPosition: {x: 0, y: 0, z: 0}
-  dragCurrentPosition: {x: 0, y: 0, z: 0}
+  cameraParentTransform: {fileID: 1457566343}
+  cameraTransform: {fileID: 512232610}
+  panLimit1: {x: 0, y: 50}
+  panLimit2: {x: 120, y: 200}
+  zoomLimitMinY: 15
+  zoomLimitMaxY: 130
+  normalSpeed: 5
+  fastSpeed: 12
+  movementSpeed: 5
+  movementTime: 5
+  rotationAmount: 5
+  zoomAmount: {x: 0, y: -10, z: 10}
 --- !u!4 &834136087
 Transform:
   m_ObjectHideFlags: 0
@@ -452,7 +404,6 @@ GameObject:
   serializedVersion: 6
   m_Component:
   - component: {fileID: 1457566343}
-  - component: {fileID: 1457566344}
   m_Layer: 0
   m_Name: Rendering
   m_TagString: Untagged
@@ -476,36 +427,6 @@ Transform:
   m_Father: {fileID: 0}
   m_RootOrder: 0
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
---- !u!114 &1457566344
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 1457566342}
-  m_Enabled: 0
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 16d7ee547532bfc4cbaf9f7ebeacead1, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  cameraTrasnform: {fileID: 512232610}
-  panLimit1: {x: 0, y: -120}
-  panLimit2: {x: 120, y: 50}
-  minY: 40
-  maxY: 150
-  normalSpeed: 2
-  fastSpeed: 5
-  movementSpeed: 1
-  movementTime: 5
-  rotationAmount: 1
-  zoomAmount: {x: 0, y: -10, z: -10}
-  newPosition: {x: 0, y: 0, z: 0}
-  newRotation: {x: 0, y: 0, z: 0, w: 0}
-  newZoom: {x: 0, y: 0, z: 0}
-  dragStartPosition: {x: 0, y: 0, z: 0}
-  dragCurrentPosition: {x: 0, y: 0, z: 0}
-  rotateStartPosition: {x: 0, y: 0, z: 0}
-  rotateCurrentPosition: {x: 0, y: 0, z: 0}
 --- !u!1 &1589499993
 GameObject:
   m_ObjectHideFlags: 0
