+++ Assets/Scripts/MouseManager.cs
@@ -1,48 +1,59 @@
-﻿
+﻿using System.Collections;
+using System.Collections.Generic;
 using UnityEngine;
 
 public class MouseManager : MonoBehaviour
 {
+    /*
+     *  This implements an orbit camera.
+     *  So in our Unity Project we take the "Rendering" component as Parent, 
+     *  which is moved arround on the x-z plane and the camera follows.
+     */
+    public Transform cameraParentTransform;
+    public Transform cameraTransform;
+
     public Vector2 panLimit1;
     public Vector2 panLimit2;
-    public float minY = 40f;
-    public float maxY = 150f;
-    public float scrollSpeed = 10f;
+    public float zoomLimitMinY;
+    public float zoomLimitMaxY;
 
-    public float panSpeed = 30f;
+    public float normalSpeed;
+    public float fastSpeed;
+    public float movementSpeed;
+    public float movementTime;
+    public float rotationAmount;
+    public Vector3 zoomAmount;
 
-    public Vector3 dragStartPosition;
-    public Vector3 dragCurrentPosition;
+    Vector3 newPosition;
+    Quaternion newRotation;
+    Vector3 newZoom;
 
+    Vector3 dragStartPosition;
+    Vector3 dragCurrentPosition;
+    Vector3 rotateStartPosition;
+    Vector3 rotateCurrentPosition;
 
 
     // Start is called before the first frame update
     void Start()
     {
-
-
+        newPosition = cameraParentTransform.position;
+        newRotation = cameraParentTransform.rotation;
+        newZoom = cameraTransform.localPosition;
     }
 
     // Update is called once per frame
     void Update()
     {
-        Vector3 pos = transform.position;
+        HandleMovementInput();
+        HandleMouseInput();
+    }
 
-        if (Input.GetKey("w") || Input.GetKey(KeyCode.UpArrow))
+    void HandleMouseInput()
+    {
+        if (Input.mouseScrollDelta.y != 0)
         {
-            pos.z += panSpeed * Time.deltaTime;
-        }
-        if (Input.GetKey("s") || Input.GetKey(KeyCode.DownArrow))
-        {
-            pos.z -= panSpeed * Time.deltaTime;
-        }
-        if (Input.GetKey("d") || Input.GetKey(KeyCode.RightArrow))
-        {
-            pos.x += panSpeed * Time.deltaTime;
-        }
-        if (Input.GetKey("a") || Input.GetKey(KeyCode.LeftArrow))
-        {
-            pos.x -= panSpeed * Time.deltaTime;
+            newZoom += Input.mouseScrollDelta.y * zoomAmount;
         }
         if (Input.GetMouseButtonDown(1))
         {
@@ -69,19 +80,87 @@ public class MouseManager : MonoBehaviour
             {
                 dragCurrentPosition = ray.GetPoint(entry);
 
-                pos = transform.position + dragStartPosition - dragCurrentPosition;
+                newPosition = cameraParentTransform.position + dragStartPosition - dragCurrentPosition;
 
             }
         }
-        float scroll = Input.GetAxis("Mouse ScrollWheel");
-        pos.y -= scroll * scrollSpeed * 100f * Time.deltaTime;
+        if (Input.GetMouseButtonDown(2))
+        {
+            rotateStartPosition = Input.mousePosition;
 
-        pos.x = Mathf.Clamp(pos.x, panLimit1.x, panLimit2.x);
-        pos.y = Mathf.Clamp(pos.y, minY, maxY);
-        pos.z = Mathf.Clamp(pos.z, panLimit1.y, panLimit2.y);
+        }
+        if (Input.GetMouseButton(2))
+        {
+            rotateCurrentPosition = Input.mousePosition;
 
-        transform.position = pos;
+            Vector3 difference = rotateStartPosition - rotateCurrentPosition;
 
+            rotateStartPosition = rotateCurrentPosition;
+
+            newRotation *= Quaternion.Euler(Vector3.up * (-difference.x / 5f));
+        }
     }
 
+    void HandleMovementInput()
+    {
+        if (Input.GetKey(KeyCode.LeftShift))
+        {
+            movementSpeed = fastSpeed;
+        }
+        else
+        {
+            movementSpeed = normalSpeed;
+        }
+        if (Input.GetKey(KeyCode.W) || Input.GetKey(KeyCode.UpArrow))
+        {
+            newPosition += (cameraParentTransform.forward * movementSpeed);
+        }
+        if (Input.GetKey(KeyCode.S) || Input.GetKey(KeyCode.DownArrow))
+        {
+            newPosition += (cameraParentTransform.forward * -movementSpeed);
+        }
+        if (Input.GetKey(KeyCode.D) || Input.GetKey(KeyCode.RightArrow))
+        {
+            newPosition += (cameraParentTransform.right * movementSpeed);
+        }
+        if (Input.GetKey(KeyCode.A) || Input.GetKey(KeyCode.LeftArrow))
+        {
+            newPosition += (cameraParentTransform.right * -movementSpeed);
+        }
+
+        if (Input.GetKey(KeyCode.Q))
+        {
+            newRotation *= Quaternion.Euler(Vector3.up * rotationAmount);
+        }
+        if (Input.GetKey(KeyCode.E))
+        {
+            newRotation *= Quaternion.Euler(Vector3.up * -rotationAmount);
+        }
+        if (Input.GetKey(KeyCode.R))
+        {
+            newZoom += zoomAmount;
+        }
+        if (Input.GetKey(KeyCode.T))
+        {
+            newZoom -= zoomAmount;
+        }
+
+        // camera parent movement:
+        // lock camera parent to boundaries
+        newPosition.x = Mathf.Clamp(newPosition.x, panLimit1.x, panLimit2.x);
+        newPosition.z = Mathf.Clamp(newPosition.z, panLimit1.y, panLimit2.y);
+        // move and rotate camera parent
+        cameraParentTransform.position = Vector3.Lerp(cameraParentTransform.position, newPosition, Time.deltaTime * movementTime);
+        cameraParentTransform.rotation = Quaternion.Lerp(cameraParentTransform.rotation, newRotation, Time.deltaTime * movementTime);
+
+
+        // lock zoom to boundaries
+        if ((newZoom.y < zoomLimitMinY) || (newZoom.y > zoomLimitMaxY))
+        {
+            newZoom = cameraTransform.localPosition;
+        }
+        // move camera for zoom
+        cameraTransform.localPosition = Vector3.Lerp(cameraTransform.localPosition, newZoom, Time.deltaTime * movementTime);
+
+    }
 }
