--- Assets/Scripts/MouseManager1.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 16d7ee547532bfc4cbaf9f7ebeacead1
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
