--- Assets/Scripts/MouseManager2.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: b1a25d9d8eada7844a2f46969fc96080
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
